use mahout and add contentbasedserivce, userbasedservice

+++ src/main/java/com/sun/movie/dal/entity/MoviePreference.java
@@ -5,9 +5,9 @@ import lombok.Data;
 
 @Data
 public class MoviePreference extends Base {
-    private int id;
-    private int user_id;
-    private int movie_id;
+    private long id;
+    private long user_id;
+    private long movie_id;
     private int preference;
     private int timestamp;
 }
